# src/secrets-service/openapi/openapi-3.0.yaml
openapi: 3.0.3
info:
    title: Verself Secrets Service API
    version: dev
servers:
    - url: https://secrets.api.anveio.com
paths:
...
    /api/v1/secrets:
        get:
            summary: List retrievable secret metadata
            operationId: list-secrets
            parameters:
                - name: limit
                  in: query
                  explode: false
                  schema:
                    type: integer
                    format: int64
                    minimum: 1
                    maximum: 200
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SecretsDTO'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
            security:
                - bearerAuth: []
            x-verself-iam:
                action: list
                audit_event: secrets.secret.list
                billing_product_id: secrets
                billing_sku_id: secrets_kv_operation
                data_classification: secret
                event_category: secrets
                openbao_role: secrets-direct-list-secrets
                operation_display: list secrets
                operation_type: read
                org_scope: token_org_id
                permission: secrets:secret:list
                rate_limit_class: read
                resource: secret
                risk_level: high
                source_product_area: Secrets
    /api/v1/secrets/{name}:
        get:
            summary: Resolve and read a retrievable secret
            operationId: read-secret
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
                    minLength: 1
                    maxLength: 255
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SecretValueDTO'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
            security:
                - bearerAuth: []
            x-verself-iam:
                action: read
                audit_event: secrets.secret.read
                billing_product_id: secrets
                billing_sku_id: secrets_kv_operation
                data_classification: secret
                event_category: secrets
                openbao_role: secrets-direct-read-secret
                operation_display: read secret
                operation_type: read
                org_scope: token_org_id
                permission: secrets:secret:read
                rate_limit_class: read
                resource: secret
                risk_level: critical
                source_product_area: Secrets
        put:
            summary: Create or rotate a retrievable secret
            operationId: put-secret
            parameters:
                - name: Idempotency-Key
                  in: header
                  description: Stable caller-provided key used to make this mutation retry-safe.
                  required: true
                  schema:
                    type: string
                    minLength: 1
                    maxLength: 128
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
                    minLength: 1
                    maxLength: 255
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PutSecretBody'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SecretDTO'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
            security:
                - bearerAuth: []
            x-verself-iam:
                action: write
                audit_event: secrets.secret.write
                billing_product_id: secrets
                billing_sku_id: secrets_kv_operation
                data_classification: secret
                event_category: secrets
                openbao_role: secrets-direct-put-secret
                operation_display: put secret
                operation_type: write
                org_scope: token_org_id
                permission: secrets:secret:write
                rate_limit_class: secret_mutation
                resource: secret
                risk_level: high
                source_product_area: Secrets
        delete:
            summary: Soft-delete a retrievable secret
            operationId: delete-secret
            parameters:
                - name: Idempotency-Key
                  in: header
                  description: Stable caller-provided key used to make this mutation retry-safe.
                  required: true
                  schema:
                    type: string
                    minLength: 1
                    maxLength: 128
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
                    minLength: 1
                    maxLength: 255
                - name: scope_level
                  in: query
                  explode: false
                  schema:
                    type: string
                    enum:
                        - org
                        - source
                        - environment
                        - branch
                - name: source_id
                  in: query
                  explode: false
                  schema:
                    type: string
                    maxLength: 255
                - name: env_id
                  in: query
                  explode: false
                  schema:
                    type: string
                    maxLength: 255
                - name: branch
                  in: query
                  explode: false
                  schema:
                    type: string
                    maxLength: 255
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteSecretInputBody'
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SecretDTO'
                default:
                    description: Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/ErrorModel'
            security:
                - bearerAuth: []
            x-verself-iam:
                action: delete
                audit_event: secrets.secret.delete
                billing_product_id: secrets
                billing_sku_id: secrets_kv_operation
                data_classification: secret
                event_category: secrets
                openbao_role: secrets-direct-delete-secret
                operation_display: delete secret
                operation_type: delete
                org_scope: token_org_id
                permission: secrets:secret:delete
                rate_limit_class: secret_mutation
                resource: secret
                risk_level: high
                source_product_area: Secrets
...
components:
    schemas:
...
        DeleteSecretInputBody:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://secrets.api.anveio.com/schemas/DeleteSecretInputBody.json
                    readOnly: true
...
        ErrorDetail:
            type: object
            additionalProperties: false
            properties:
                location:
                    type: string
                    description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
                message:
                    type: string
                    description: Error message text
                value:
                    description: The value at the given location
        ErrorModel:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://secrets.api.anveio.com/schemas/ErrorModel.json
                    readOnly: true
                detail:
                    type: string
                    description: A human-readable explanation specific to this occurrence of the problem.
                    examples:
                        - Property foo is required but is missing.
                errors:
                    type: array
                    description: Optional list of individual error details
                    items:
                        $ref: '#/components/schemas/ErrorDetail'
                instance:
                    type: string
                    description: A URI reference that identifies the specific occurrence of the problem.
                    format: uri
                    examples:
                        - https://example.com/error-log/abc123
                status:
                    type: integer
                    description: HTTP status code
                    format: int64
                    examples:
                        - 400
                    maximum: 599
                title:
                    type: string
                    description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
                    examples:
                        - Bad Request
                type:
                    type: string
                    description: A URI reference to human-readable documentation for the error.
                    format: uri
                    default: about:blank
                    examples:
                        - https://example.com/errors/example
...
        PutSecretBody:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://secrets.api.anveio.com/schemas/PutSecretBody.json
                    readOnly: true
                branch:
                    type: string
                    maxLength: 255
                env_id:
                    type: string
                    maxLength: 255
                scope_level:
                    type: string
                    enum:
                        - org
                        - source
                        - environment
                        - branch
                source_id:
                    type: string
                    maxLength: 255
                value:
                    type: string
                    maxLength: 65536
            required:
                - value
...
        SecretDTO:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://secrets.api.anveio.com/schemas/SecretDTO.json
                    readOnly: true
                branch:
                    type: string
                created_at:
                    type: string
                    format: date-time
                current_version:
                    type: string
                env_id:
                    type: string
                kind:
                    type: string
                name:
                    type: string
                scope_level:
                    type: string
                secret_id:
                    type: string
                source_id:
                    type: string
                updated_at:
                    type: string
                    format: date-time
            required:
                - secret_id
                - kind
                - name
                - scope_level
                - current_version
                - created_at
                - updated_at
        SecretValueDTO:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://secrets.api.anveio.com/schemas/SecretValueDTO.json
                    readOnly: true
                branch:
                    type: string
                created_at:
                    type: string
                    format: date-time
                current_version:
                    type: string
                env_id:
                    type: string
                kind:
                    type: string
                name:
                    type: string
                scope_level:
                    type: string
                secret_id:
                    type: string
                source_id:
                    type: string
                updated_at:
                    type: string
                    format: date-time
                value:
                    type: string
            required:
                - value
                - secret_id
                - kind
                - name
                - scope_level
                - current_version
                - created_at
                - updated_at
        SecretsDTO:
            type: object
            additionalProperties: false
            properties:
                $schema:
                    type: string
                    description: A URL to the JSON Schema for this object.
                    format: uri
                    examples:
                        - https://secrets.api.anveio.com/schemas/SecretsDTO.json
                    readOnly: true
                secrets:
                    type: array
                    items:
                        $ref: '#/components/schemas/SecretDTO'
            required:
                - secrets
...
    securitySchemes:
        bearerAuth:
            type: http
            description: Zitadel-issued bearer token scoped to the secrets-service API audience.
            scheme: bearer
            bearerFormat: JWT